frontend-gradio/requirements.txt
gradio>=4.19.2
markdown2>=2.4.12
pytest>=8.0.2
python-dotenv>=1.0.1